HSUI-32 clean up table
