+++ packages/hs-react-ui/src/components/Table/TableStyledComponents.tsx
@@ -1,0 +1,117 @@
+import styled from 'styled-components';
+import Icon from '@mdi/react';
+import colors from '../../enums/colors';
+import fonts from '../../enums/fonts';
+import { RowProps } from './Table';
+
+export const TableContainer = styled.table`
+  ${({ reachedMinWidth }: { reachedMinWidth?: boolean }) => `
+    width: ${reachedMinWidth ? '100%' : 'auto'};
+    ${fonts.body}
+    background-color: ${colors.background};
+    border-collapse: collapse;
+
+    border-radius: 8px;
+    overflow: hidden;
+  `}
+`;
+
+export const Header = styled.tr`
+  ${({ columnGap, columnWidths }: RowProps) => `
+    display: grid;
+    grid-template-columns: ${columnWidths};
+    padding: 0rem 2rem;
+    column-gap: ${columnGap};
+    user-select: none;
+
+    background-color: ${colors.primary};
+    color: white;
+  `}
+`;
+
+export const HeaderCell = styled.th`
+  ${({ sortable }: { sortable: boolean }) => `
+    display: flex;
+    flex-flow: row;
+    cursor: pointer;
+    padding: 1rem 0rem 1rem 1rem;
+    margin-left: -1rem;
+
+    transition: background-color 0.5s;
+
+    ${sortable ? '' : 'pointer-events: none;'}
+
+    &:hover {
+      background-color: rgba(255, 255, 255, 0.2);
+    }
+  `}
+`;
+
+export const ResponsiveTitle = styled.span`
+  ${({ sortable }: { sortable: boolean }) => `
+    color: ${colors.primary};
+    padding: 0.5rem;
+    user-select: none;
+    cursor: pointer;
+    margin-right: .5rem;
+    background-color: rgba(0,0,0,0.05);
+    border-radius: .5rem;
+    ${sortable ? '' : 'pointer-events: none;'}
+  `}
+`;
+
+export const Row = styled.tr`
+  ${({ columnGap, columnWidths, reachedMinWidth, isCollapsed = false }: RowProps) => `
+    display: ${isCollapsed ? 'none' : 'grid'};
+    grid-template-columns: ${reachedMinWidth ? '100%' : columnWidths};
+    padding: ${reachedMinWidth ? '1rem' : '0rem'} 2rem;
+    row-gap: .5rem;
+    column-gap: ${columnGap};
+    position: relative;
+    background-color: white;
+
+    &:not(:last-child) {
+      border-bottom: 1px solid rgb(211, 214, 215);
+    }
+
+    &:before {
+      content: '';
+      z-index: 0;
+      position: absolute;
+      top: 0; left: 0;
+      width: 100%;
+      height: 100%;
+      background-color: rgba(0,0,0,0.2);
+      opacity: 0;
+      transition: opacity .3s;
+
+      pointer-events: none;
+    }
+    &:hover:before {
+      opacity: .3;
+    }
+  `}
+`;
+
+export const GroupRow = styled(Row)`
+  background-color: #e8e8e8;
+`;
+
+export const Cell = styled.td`
+  display: block;
+  padding: 1rem 0rem;
+  word-break: break-word;
+  hyphens: auto;
+  width: unset;
+`;
+
+export const SortIcon = styled(Icon)`
+  ${({ direction }: { direction?: boolean | null }) => `
+    margin-left: 1rem;
+    fill: white;
+    width: 1rem;
+    transition: transform .2s, opacity .5s;
+    opacity: ${direction === null ? 0 : 1};
+    transform: rotate(${direction ? 0 : 180}deg);
+  `}
+`;
